Websocket - Fetch Kick Channel & Chat Data LIVE
This is a simple javascript code for a node.js websocket, which returns multiple data from a Kick Channel.

The `fetchData(channelId)` function returns an object containing various data related to the specified channel on the website (`https://kick.com`). Here are the properties that the returned object contains:

1. `followersCount`: The number of followers for the channel.
2. `livestreamStatus`: A string indicating the livestream status of the channel. It will be either "Online" if the channel is currently live streaming or "Offline" if not.
3. `viewersCount`: The number of viewers currently watching the livestream (if the channel is online).
4. `messages`: An array containing the messages from the channel. Each message in the array may have various properties such as the message text, timestamp, sender, etc.
5. `profilePic`: The URL of the channel owner's profile picture.
6. `followersOnly`: A boolean value indicating whether the chatroom is set to be accessible only to followers (`true`) or not (`false`).
7. `subscriberOnly`: A boolean value indicating whether the chatroom is set to be accessible only to subscribers (`true`) or not (`false`).

For example, if you call `fetchData('xqc')`, it will fetch data from the channel with ID 'xqc' on `https://kick.com` and return an object with the properties mentioned above, providing information about the channel's followers, livestream status, viewers, chatroom settings, and messages. If an error occurs during the data fetching process, the function will return `null`, and an error message will be logged to the console.